Reordering to make it more readable
